Change ID data type

+++ src/sql/create-tables.sql
@@ -9,7 +9,7 @@ GO
 
 CREATE TABLE [dbo].[update_packages]
 (
-    [id]                [bigint] IDENTITY(1,1) NOT NULL,  -- The update package's ID.
+    [id]                [int] IDENTITY(1,1) NOT NULL,  -- The update package's ID.
     [file_name]         [nvarchar](1024) NOT NULL,        -- The update package's file name.
     [file_hash_sha1]    [binary](20) NOT NULL             -- The update package's file hash that computed with SHA1.
 )
@@ -19,7 +19,7 @@ GO
 
 CREATE TABLE [dbo].[update_modules]
 (
-    [id]                [bigint] IDENTITY(1,1) NOT NULL,  -- The update module's ID.
+    [id]                [int] IDENTITY(1,1) NOT NULL,  -- The update module's ID.
     [file_name]         [nvarchar](1024) NOT NULL,        -- The update module's file name.
     [file_hash_sha1]    [binary](20) NOT NULL             -- The update module's file hash that computed with SHA1.
 )
